+++ CHANGELOG.md
@@ -34,6 +34,7 @@
     - Fixed MSLE computation in Residual Analysis report
     - Renamed RMSError Metric to RMSE
     - Embedders no longer implement Estimator interface
+    - Added error statistics to Residual Analysis report
 
 - 0.0.11-beta
     - K Means now uses mini batch GD instead of SGD